+++ lib/karafka/web/deserializer.rb
@@ -6,12 +6,21 @@ module Karafka
     #
     # @note We use `symbolize_names` because we want to use the same convention of hash building
     #   for producing, consuming and displaying metrics related data
+    #
+    # @note We have to check if we compress the data, because older Web-UI versions were not
+    #   compressing the payload.
     class Deserializer
       # @param message [::Karafka::Messages::Message]
       # @return [Object] deserialized data
       def call(message)
+        raw_payload = if message.headers.key?('zlib')
+          Zlib::Inflate.inflate(message.raw_payload)
+        else
+          message.raw_payload
+        end
+
         ::JSON.parse(
-          message.raw_payload,
+          raw_payload,
           symbolize_names: true
         )
       end